% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{a_dcc_loglik}
\alias{a_dcc_loglik}
\title{A-DCC log-likelihood (second step)}
\usage{
a_dcc_loglik(param, res, K_c = NULL)
}
\arguments{
\item{param}{Vector of starting values.}

\item{res}{Array of standardized daily returns, coming from the first step estimation.}

\item{K_c}{\strong{optional} Number of initial observations to exclude from the estimation}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{t}.
}
\description{
Obtains the log-likelihood of the A-DCC model in the second step.
For details, see \insertCite{cappiello2006asymmetric;textual}{dccmidas} and \insertCite{dcc_engle_2002;textual}{dccmidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
