% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{a_dcc_mat_est}
\alias{a_dcc_mat_est}
\title{Obtains the matrix H_t and R_t, under the A-DCC model}
\usage{
a_dcc_mat_est(est_param, res, Dt, K_c = NULL)
}
\arguments{
\item{est_param}{Vector of estimated values}

\item{res}{Array of standardized daily returns, coming from the first step estimation}

\item{Dt}{Diagonal matrix of standard deviations}

\item{K_c}{\strong{optional} Number of initial observations to exclude from the H_t and R_t calculation}
}
\value{
A list with the \eqn{H_t} and \eqn{R_t} matrices, for each \eqn{t}.
}
\description{
Obtains the matrix H_t and R_t, under the A-DCC model
For details, see \insertCite{cappiello2006asymmetric;textual}{dccmidas} and \insertCite{dcc_engle_2002;textual}{dccmidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
