% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{a_dccmidas_loglik}
\alias{a_dccmidas_loglik}
\title{A-DCC-MIDAS log-likelihood (second step)}
\usage{
a_dccmidas_loglik(param, res, lag_fun = "Beta", N_c, K_c)
}
\arguments{
\item{param}{Vector of starting values.}

\item{res}{Array of standardized daily returns, coming from the first step estimation.}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{t}.
}
\description{
Obtains the log-likelihood of the A-DCC-MIDAS model in the second step.
For details, see \insertCite{cappiello2006asymmetric;textual}{dccmidas} and \insertCite{dcc_engle_2002;textual}{dccmidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
