% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dBEKK_mat_est}
\alias{dBEKK_mat_est}
\title{dBEKK covariance matrix}
\usage{
dBEKK_mat_est(param, ret)
}
\arguments{
\item{param}{Vector of starting values.}

\item{ret}{Txk matrix of daily returns. At the moment, k can be at most 4}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{t}.
}
\description{
Obtains the conditional covariance matrix from the diagonal BEKK model.
}
\references{
\insertAllCited{}
}
\keyword{internal}
