% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deco_loglik}
\alias{deco_loglik}
\title{DECO log-likelihood (second step)}
\usage{
deco_loglik(param, res, K_c = NULL)
}
\arguments{
\item{param}{Vector of starting values.}

\item{res}{Array of standardized daily returns, coming from the first step estimation.}

\item{K_c}{\strong{optional} Number of initial observations to exclude from the estimation}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{t}.
}
\description{
Obtains the log-likelihood of the DECO models in the second step.
For details, see \insertCite{engle2012dynamic;textual}{dccmidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
