% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deco_mat_est}
\alias{deco_mat_est}
\title{Obtains the matrix H_t and R_t, under the DECO model}
\usage{
deco_mat_est(est_param, res, Dt, K_c = NULL)
}
\arguments{
\item{res}{Array of standardized daily returns, coming from the first step estimation.}

\item{K_c}{\strong{optional} Number of initial observations to exclude from the H_t and R_t calculation}

\item{param}{Vector of estimated values.}
}
\value{
A list with the \eqn{H_t} and \eqn{R_t} matrices, for each \eqn{t}.
}
\description{
Obtains the matrix H_t and R_t, under the DECO model
For details, see \insertCite{engle2012dynamic;textual}{dccmidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
