% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftse100}
\alias{ftse100}
\title{FTSE 100 data}
\format{
An object of class \code{"xts"}.
}
\source{
Realized library of the \href{https://realized.oxford-man.ox.ac.uk/data/download}{Oxford-Man Institute}
}
\usage{
data(ftse100)
}
\description{
Daily data on FTSE 100 collected from the realized library of
the Oxford-Man Institute \insertCite{heber_2009}{dccmidas}.
}
\details{
ftse100 includes the open price (open_price), the realized variance
(rv5), and the close price (close_price). The realized variance has been calculated
using intradaily intervals of five minutes \insertCite{andersen_boll_1998}{dccmidas}.
}
\examples{
head(ftse100)
summary(ftse100)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
