% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%^\%}
\alias{\%^\%}
\title{Power of a matrix}
\usage{
x \%^\% p
}
\arguments{
\item{x}{A square matrix}

\item{p}{The power of interest}
}
\value{
The resulting matrix x^(p)
}
\description{
Reports the power of a matrix.
}
\keyword{internal}
