% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_cov_est}
\alias{moving_cov_est}
\title{Obtains the matrix H_t, under the Moving Covariance model}
\usage{
moving_cov_est(rt, V)

moving_cov_est(rt, V)
}
\arguments{
\item{rt}{List of daily returns}

\item{V}{Length of the rolling window adopted}
}
\value{
A list with the \eqn{H_t} matrix, for each \eqn{t}.

A list with the \eqn{H_t} matrix, for each \eqn{t}.
}
\description{
Obtains the matrix H_t, under the Moving Covariance model.

Obtains the matrix H_t, under the Moving Covariance model.
}
\keyword{internal}
