% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nasdaq}
\alias{nasdaq}
\title{NASDAQ data}
\format{
An object of class \code{"xts"}.
}
\source{
Realized library of the \href{https://realized.oxford-man.ox.ac.uk/data/download}{Oxford-Man Institute}
}
\usage{
data(nasdaq)
}
\description{
Daily data on NASDAQ collected from the realized library of
the Oxford-Man Institute \insertCite{heber_2009}{dccmidas}.
}
\details{
nasdaq includes the open price (open_price), the realized variance
(rv5), and the close price (close_price). The realized variance has been calculated
using intradaily intervals of five minutes \insertCite{andersen_boll_1998}{dccmidas}.
}
\examples{
head(nasdaq)
summary(nasdaq)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
