% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.dccmidas}
\alias{print.dccmidas}
\title{Print method for 'dccmidas' class}
\usage{
\method{print}{dccmidas}(x, ...)
}
\arguments{
\item{x}{An object of class 'dccmidas'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
Print method for 'dccmidas' class
}
\keyword{internal}
