% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sBEKK_loglik}
\alias{sBEKK_loglik}
\title{sBEKK log-likelihood}
\usage{
sBEKK_loglik(param, ret)
}
\arguments{
\item{param}{Vector of starting values.}

\item{ret}{Txk matrix of daily returns. At the moment, k can be at most 5}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{t}.
}
\description{
Obtains the log-likelihood of the scalar BEKK model.
}
\references{
\insertAllCited{}
}
\keyword{internal}
