% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.dccmidas}
\alias{summary.dccmidas}
\title{Summary method for 'dccmidas' class}
\usage{
\method{summary}{dccmidas}(object, ...)
}
\arguments{
\item{object}{An object of class 'dccmidas', that is the result of a call to \code{\link{dcc_fit}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Returns the printed tables of the univariate and multivariate steps as well as
some additional information about the number of observations, sample period, and information criteria
}
\description{
Summary method for 'dccmidas' class
}
\keyword{internal}
